pyproject.toml
[project]
name = "marketing-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "django>=5.2.11",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "faker>=40.5.1",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pyotp>=2.9.0",
    "qrcode>=8.2",
    "redis>=7.2.0",
    "twilio>=9.10.4",
]
